refactored Base and added create for #10
