+++ test/test_scopes.py
@@ -19,7 +19,7 @@ class Watermelon(base.Base):
 
 class TestScopes(unittest.TestCase):
 
-    @mock.patch("base.query_methods.Query")
+    @mock.patch("base.Query")
     def test_defines_class_method_on_model(self, Query):
         Cantaloupe.ripe()
         Query.assert_called_with(Cantaloupe)
@@ -29,17 +29,17 @@ class TestScopes(unittest.TestCase):
     def test_class_method_has_reasonable_signature(self):
         self.assertEqual(Cantaloupe.ripe.__name__, "<scope>ripe")
 
-    @mock.patch("base.query_methods.Query.where")
+    @mock.patch("base.Query.where")
     def test_defines_method_on_classes_queries(self, where):
-        query = base.query_methods.Query(Cantaloupe)
+        query = base.Query(Cantaloupe)
         query.ripe()
         where.assert_called_with(ripe=1)
 
-    @mock.patch("base.query_methods.Query.where")
+    @mock.patch("base.Query.where")
     def test_does_not_define_method_on_other_queries(self, where):
-        query = base.query_methods.Query(Cantaloupe)
+        query = base.Query(Cantaloupe)
         query.ripe()
-        query2 = base.query_methods.Query(Watermelon)
+        query2 = base.Query(Watermelon)
         with self.assertRaises(AttributeError) as e:
             query2.ripe()
         self.assertEqual(e.exception.message,
